Remove unused internal/pem and internal/util packages. Move version to only place it's used in response
