+++ controller/response/version.go
@@ -14,4 +14,12 @@
 	limitations under the License.
 */
 
-package util
+package response
+
+func GetApiVersion() string {
+	return "0.0.1"
+}
+
+func GetApiEnrollmentVersion() string {
+	return "0.0.1"
+}
